Allow passing a file-path as PR body
Sometimes the PR body contains a lot of information. In this case
passing it as argument to the action will cause a shell error of having
an argument list that's too long. To circumvent this we provide the
option of providing a file path that is loaded by Node instead.

We prefer the file-path but fall back to the body.